(feat) Make webhook trigger clone, compile, test in pushed branch
If a branch is pushed on GitHub (in the 'ci' repo), a webhook will send
an HTTP POST request to our server (currently via ngrok, later to Axel's
Raspberry Pi) which will clone the pushed branch and use Maven to run
tests on the latest commit in that branch. The output of these tests
are posted to our MongoDB database.

Targets: #2, #3